refactor matrix context
