+++ src/components/Spreadsheet/data/constants.ts
@@ -1,0 +1,3 @@
+export const parentTag = 'TD'
+export const inputTag = 'INPUT'
+export const selected = 'selected'
