--- src/components/Spreadsheet/logic/Table.ts
@@ -1,8 +1,0 @@
-// Select cell
-// Select columns
-// Select rows
-// Selection clipboard
-// Selection fill
-// Clipboard paste
-// Drag & Drop
-// File drop fill
