+++ src/context/matrix/MatrixContext.ts
@@ -1,0 +1,4 @@
+import { MatrixContext as MatrixContextType } from '@/context/matrix/data/types'
+import { createContext } from 'react'
+
+export const MatrixContext = createContext(null as unknown as MatrixContextType)
