+++ src/context/matrix/data/types.d.ts
@@ -1,0 +1,13 @@
+export type Cell = {
+  x: number
+  y: number
+  inputValue: string
+  computedValue: string
+  update: (value: string) => void
+  id: string
+  references: string[]
+}
+export type MatrixContext = {
+  matrix: Cell[][]
+  save: () => void
+}
